MSH Project/gaji_elo/requirements.txt
# Core dependencies
streamlit==1.32.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# NLP and text processing
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.4.0

# Vector database for RAG
chromadb==0.4.22
langchain==0.1.5
langchain-community==0.0.17

# API clients
openai==1.10.0  # For OpenRouter
httpx==0.26.0
aiohttp==3.9.3

# Data validation
pydantic==2.5.3

# Visualization
plotly==5.18.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==23.12.1
flake8==7.0.0